Wrapper and merge script updates for value range bug
Because of SQLite issues, HappyFace and the merge script
were having trouble with integers reaching the type
limit with the Postgres (and possibly other) backends.

The wrapper now increases the possible values of a column
(alter table int -> bigint) if required by either change
in the "coded schema class" and if the a row to be inserted
would exceed the limit.

The system is rather automatic, not requireing intervention
by maintainers, but does "secretly" increase DB memory usage
during a run if required by the data. Our decision was that
capturing data was more important then resource usage, especially
since the increase is usually non-significant and fast.


git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/trunk@688 5ec1a27d-59bc-451c-a1e5-77859e3e35f3