Altered table adding/dropping to work with ForeignKeys
The MetaData method drop_all() and create_all() were
used with restriction to certail tables. It automatically
resolves any dependices. We had to sacrifice individual
adding and dropping, now only the whole set or nothing
is changed...

Tested to work with SQLite and Postgres (adding tables,
altering, adding, dropping columns. Dropping tables was
not tested, but should work).

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@761 5ec1a27d-59bc-451c-a1e5-77859e3e35f3