Implemented Certificate Authorization
The CherryPy tool in hf/auth.py checks for
client cert info from the host HTTP server
and checks a preloaded file with authorized
DNs and/or the return status of an script
to find out whether the user has access.

So far, the plot generator, the XML output
and the weboutput are cert aware, so basically
everything.

Formatting and details need to be improved.



git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@809 5ec1a27d-59bc-451c-a1e5-77859e3e35f3