Skip data exctraction when preparation failed
The extractData() method should not be called
when the prepareAcquisition() in the same module
failes with an exception.

Subsequently, the data from extractData() is
not passed to the database, what usually leads
to errors anyway.

Additionally, a better error output was added.
So basically the exception message is passed
from prepareAcquisition() to render().

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@867 5ec1a27d-59bc-451c-a1e5-77859e3e35f3