New caching behavior for categories
Different Expires-header for requests with/without set
date/time information.

If no explicit time is given (-> most recent run shall be displayed)
it is checked if the cached page is older than the most recent run
in the database. If that is the case, all variants of the current
URI is removed from cache (it is not possible to remove *this*
variant as far as I know).

This fixes the issue where acquisition is shadowed by caching,
so that fresh data is not available to the user.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@889 5ec1a27d-59bc-451c-a1e5-77859e3e35f3