Fix plotgenerator as mentioned in ticket #57, including more documentation
The plotgenerator was reworked so that clashes between columns and math
functions with the same name are avoided. For that, when math
expressions are used, $colname is the access to the column named
colname.

If the given expression matches a column name exactly, no math
expression is evaluated.

This shouldn't cause any problems, only the few modules using
expressions have to be updated.

For this to work, separation at commas inside the math expressions has
to be avoided in plot configuration. Therefore, commas in math
expressions have to be escaped with a leading backslash. Because they
are uncommon in math expressions anyway, there is no need for backslash
escaping in the seldom case a backslash is required before a comma
(which should be never, anyway).

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@913 5ec1a27d-59bc-451c-a1e5-77859e3e35f3