Display detailed module information in case of error
So far, the module information panel was truncated if there was either
an error or there was no data at that point in time. Because of this,
e.g. no source URL is displayed in case of a "simple" error, so
information is withdrawn from user.

This fix is made in preparation for ticket #68

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@928 5ec1a27d-59bc-451c-a1e5-77859e3e35f3