Implemented file upload interface
This new feature allows for remote entities to push files on the
HappyFace server via HTTP. The upload is done via a "file form",
HF then saves the uploaded file in the specified directory.

For security reasons, access is restricted. So far, only HTTP Basic Auth
is supported, though HTTP Digest should be easy with the CP interfaces.
Also, the cert auth infrastructure available in HF could be utilized.

The documentation needs to be extended, yet.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@942 5ec1a27d-59bc-451c-a1e5-77859e3e35f3