source url bugfix: extractData exception caused url to be not saved
Because the source_url was calculated after the call to extractData()
within the same try-block, an exception in the module code caused the
source url to be not displayed, contrary to the goal of the new concpt.

The code was moved to the upper level try-block, before the call to
extractData().

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@946 5ec1a27d-59bc-451c-a1e5-77859e3e35f3