Added CSS switch feature
To fix the second part of ticket #74, the main CSS can be replaced by
a mobile optimized stylesheet.

A cookie "css" is set on the device containing the URL to the CSS file to
use. The value of the cookie can be set via the css GET parameter, so
appending e.g ?css=/Happy/Face/URL/static/css/hf.css one can force usage
of the default CSS file. An empty path causes the cookie to be deleted
(Expires=0) and to have HF choose a stylesheet automaticaly.

On each page request, a handler checks for the presence of a CSS cookie and
uses its contents as cookie URL. If no cookie is set or if the path is
empty, the user-agent string is compared to a fixed list of fragments. Each
fragment is characteristic for certain mobile device browsers, so we can
decide if we want to use the default CSS or a mobile CSS.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@951 5ec1a27d-59bc-451c-a1e5-77859e3e35f3