Created Smartphone optimized, responsive CSS
The new layout constist of three responsive views:
A: min-width < 480px
B: 480px < min-width < 1024px
C: 1024px < min-width

View A is the smallest view, the header and fastnav
are not fixed to the viewport, they are simply placed
at the beginning of the page. The fastnav entries are
displayed in-line, without the text.

View B has the fastnav on the side, but only images
are shown.

View C is the classic view, where the header "stays on top" if
you scroll down the content. The main difference is, that the
fastnav is always expanded, but has a fixed width and does only
show troubleing modules.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@954 5ec1a27d-59bc-451c-a1e5-77859e3e35f3