Added legacy HF3 stylesheet
The legacy.css can be used if the "fancy" media-query
responsive, smartphone optimized hf.css is not wanted.
This can be appropriate if HF is aimed for use with
older browsers.

It is merely the old CSS before the smartphone updates,
modified for use with the new category nav template,
lacking any support for the new fastnav template.

To use it, just set it with hf_default_css in [paths] or
select it per user with the ?css= switching feature (using cookies).

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@959 5ec1a27d-59bc-451c-a1e5-77859e3e35f3