Implemented module short_name system
A module can now have a short_name, too. The short name can be
optionally specified in the configuration to avoid line wrapping
the the fastnav bar if the normal module name is very long.

Documentation and templates were updated. The short name is used
in the fastnavbar and if it was explicitely specified in the config,
it will be shown in the module panel.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@962 5ec1a27d-59bc-451c-a1e5-77859e3e35f3