defaultconfig, DownloadService: Export URL of tmp dir
The tmp dir can be user accessible, as it is in the default config.
So far, the tmp URL was not specified in the default config.

Now, the DownloadFile uses the tmp_url config option to export the
HTTP accessible URL of a file, if a tmp_url was specified. If not,
the path is exported.

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@975 5ec1a27d-59bc-451c-a1e5-77859e3e35f3