Module System: Extended error string by downloaded files
ModuleProxy was extended so that the error string stored in the database
contains a list of the files downloaded for this module, if any. The actual
message and each file specification are sperataded by null-bytes (\0), each
file is a |||-separated tripplet (source URL, filename, URL).

In case something goes wrong during data acquisition, ModuleProxy instructs
the downloadService to keep the files associated with the specific module.

ModuleBase decodes the error string in the constructor, so the error_string
attribute contains the usual information. A new downloaded_files attribute
has been added, containing a list of tripplets. The list is *only* populated
in case of an error!

git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/HappyFace/branches/v3.0@977 5ec1a27d-59bc-451c-a1e5-77859e3e35f3